+++ src/main/java/me/lambdaurora/lambdynlights/mixin/lightsource/LivingEntityMixin.java
@@ -37,7 +37,7 @@ public abstract class LivingEntityMixin extends Entity implements DynamicLightSo
         } else {
             int luminance = 0;
             for (ItemStack equipped : this.getItemsEquipped()) {
-                luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(equipped));
+                luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(equipped, this.isSubmergedInWater()));
             }
 
             this.lambdynlights_luminance = luminance;