+++ src/main/java/me/lambdaurora/lambdynlights/mixin/ltr/LilTaterBlocksMixin.java
@@ -11,15 +11,18 @@ package me.lambdaurora.lambdynlights.mixin.ltr;
 
 import me.lambdaurora.lambdynlights.LambDynLights;
 import me.lambdaurora.lambdynlights.api.DynamicLightHandlers;
+import me.lambdaurora.lambdynlights.util.LilTaterBlockEntityAccessor;
 import net.minecraft.block.entity.BlockEntityType;
 import net.minecraft.item.ItemStack;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Pseudo;
 import org.spongepowered.asm.mixin.gen.Accessor;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(targets = "mods.ltr.registry.LilTaterBlocks")
+@Pseudo
+@Mixin(targets = "mods.ltr.registry.LilTaterBlocks", remap = false)
 public class LilTaterBlocksMixin
 {
     @Accessor("LIL_TATER_BLOCK_ENTITY")
@@ -35,7 +38,7 @@ public class LilTaterBlocksMixin
             int luminance = 0;
             if (!((LilTaterBlockEntityAccessor) entity).lambdynlights_isEmpty()) {
                 for (ItemStack item : ((LilTaterBlockEntityAccessor) entity).lambdynlights_getItems()) {
-                    luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(item));
+                    luminance = Math.max(luminance, LambDynLights.getLuminanceFromItemStack(item, !entity.getCachedState().getFluidState().isEmpty()));
                 }
             }
             return luminance;